Move rollup config around
